chore(content): fix and update preparemessage function
