+++ src/message.ts
@@ -1,0 +1,23 @@
+import { Config } from "./Config";
+import { Context } from "semantic-release";
+
+export function preparemessage(config: Config, context: Context) {
+  const { env, nextRelease, logger } = context;
+  logger.log('prepare message.');
+  if (!env.DISCORD_CUSTOM_MESSAGE || !config.custom_message) {
+    logger.log('prepare default message.');
+    const message = {
+      text: `The :rocket: ${nextRelease.type} version "${nextRelease.version}" :heavy_check_mark: has been released.\n\n${nextRelease.notes}`,
+      username: config.username
+    }
+    return message;
+  }
+  else {
+    logger.log('prepare custom message.');
+    const message = {
+      text: `${config.custom_message}`,
+      username: config.username
+    }
+    return message;
+  }
+}
